Refactor and logic changed for kakao alert

+++ Kakao_test/kakao_project_module/src/main/java/com/kakao/service/ProjectKakaoServiceImpl.java
@@ -21,7 +21,7 @@ public class ProjectKakaoServiceImpl implements ProjectKakaoService {
     private final KakaoTemplateManager kakaoTemplateManager;
 
     @Override
-    public ReturnDto sendKakaoMessage(Long memberId, Long templateId) {
+    public ReturnDto sendKakaoMessage(Long memberId, int templateId) {
         ReturnDto messageResult = kakaoSendManager.sendMessage(memberId, templateId);
         return messageResult;
     }
