+++ Kakao_test/kakao_project_module/src/main/java/com/kakao/service/contract/ProjectKakaoService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface ProjectKakaoService {
 
     //send message
-    ReturnDto sendKakaoMessage (Long memberId, Long templateId);
+    ReturnDto sendKakaoMessage (Long memberId, int templateId);
 
     //manage templates
